pyproject.toml
[tool.poetry]
name = "elp-ai-assistant"
version = "0.1.0"
description = "Projeto de IA para ELP da MTI"
authors = ["LoboProgrammingg <matheusloboo2001@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/LoboProgrammingg/IA-ELP-2026.git"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
google-generativeai = "^0.8.5"
faiss-cpu = "^1.8.0"
pandas = "^2.2.0"
docling = "^2.54.0"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"
tavily-python = "^0.7.12"
agno = "^0.1.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
